Refactor streaming module imports to fix cyclic import
