+++ app/data_layer/streaming/__init__.py
@@ -1,2 +1,1 @@
-from .streamer import Streamer  # isort:skip
 from .streamers import *
