+++ app/data_layer/streaming/streamers/kafka_streamer.py
@@ -4,7 +4,7 @@ from typing import Optional
 from kafka import KafkaProducer
 from omegaconf import DictConfig
 
-from app.data_layer.streaming import Streamer
+from app.data_layer.streaming.streamer import Streamer
 from app.utils.common.logger import get_logger
 
 logger = get_logger(Path(__file__).name)
