+++ app/sockets/connections/smartsocket_connection.py
@@ -7,7 +7,7 @@ from omegaconf import DictConfig
 from app.data_layer.database.sqlite.crud.smartapi_crud import (
     get_smartapi_tokens_by_all_conditions,
 )
-from app.data_layer.streaming import Streamer
+from app.data_layer.streaming.streamer import Streamer
 from app.sockets.connections.websocket_connection import WebsocketConnection
 from app.sockets.twisted_sockets import SmartSocket
 from app.utils.common import init_from_cfg
